Fixed multithreading crash and fixed warnings
If 2 events for generating PerlinNoise arrived on one frame, threads are
not yet initialized, so they cannot be joined. Now this situation is
processed correctly

Added ivec2, split math.hpp header into several files

Removed several implicit casts

+++ visual/ecs/camera_module.hpp
@@ -6,7 +6,7 @@
 
 struct CameraState {
   vec2 center{0.0f, 0.0f};
-  vec2 display_dimentions;
+  ivec2 display_dimentions;
   Box2 view;
 
   float zoom = 1.0f;
